create component for revenue chart
